added all tasks
