--- 4. Creating a Test File Using POM/README.md
@@ -1,3 +1,0 @@
-1. Your task is to fill in all the ... in the test_text_bike_after_click.py file. 
-   The tests should pass without errors
-2. You can find the solution in the Solutions folder. File test_text_bike_after_click.py
