+++ 4. Creating a Test File Using POM/Task1/test_text_bike_after_click.py
@@ -11,21 +11,21 @@ def test_custom_bike_option():
     # Create an instance of the page class
     urban_routes_page = UrbanRoutesPage(driver)
 
-    # Step 3: Use POM methods to perform actions on the page
+    # Step 2: Use POM methods to perform actions on the page
     # Enter "From" and "To" locations.
-    urban_routes_page.enter_from_location('East 2nd Street, 601')
-    urban_routes_page.enter_to_location('1300 1st St')
+    ...
 
     # Select the "Custom" option.
-    urban_routes_page.click_custom_option()
+    ...
     time.sleep(2)  # Adding delay for visibility; optional
 
     # Click the "Bike" icon.
-    urban_routes_page.click_bike_icon()
+    ...
     time.sleep(2)  # Adding delay for visibility; optional
 
-    # Step 4: Verify the Bike text is displayed correctly
-    actual_value = urban_routes_page.get_bike_text()
-    expected_value = "Bike"
-    assert expected_value in actual_value, f"Expected '{expected_value}', but got '{actual_value}'"
+    # Step 3: Verify the Bike text is displayed correctly
+    actual_value = ...
+    expected_value = ...
+    assert ...
     driver.quit()
+
