+++ 4. Creating a Test File Using POM/Task1/urban_routes_main_page.py
@@ -6,9 +6,9 @@ class UrbanRoutesPage:
     # Locators as class attributes
     FROM_LOCATOR = (By.ID, 'from')
     TO_LOCATOR = (By.ID, 'to')
-    CUSTOM_OPTION_LOCATOR = (By.XPATH, '//div[text()="Custom"]')
-    BIKE_ICON_LOCATOR = (By.XPATH, '//img[@src="/static/media/bike.fb41c762.svg"]')
-    BIKE_TEXT_LOCATOR = (By.XPATH, '//div[contains(text(),"Bike")]')
+    CUSTOM_OPTION_LOCATOR = ...
+    BIKE_ICON_LOCATOR = ...
+    BIKE_TEXT_LOCATOR = ...
 
     def __init__(self, driver):
         self.driver = driver  # Initialize the driver
