--- 6. Setup_class() and teardown_class() methods/README.md
@@ -1,3 +1,0 @@
-1. Your task is to fill in all the ... in the test_setup_teardown_urban_routes.py file. 
-   The tests should pass without errors
-2. You can find the solution in the Solutions folder. File test_setup_teardown_urban_routes.py
