+++ README.md
@@ -1,9 +1,10 @@
+To install virtual environment
+
 1. Run in the terminal  
 python -m venv .venv
 2. Windows  
 .venv\Scripts\activate  
    MAC  
-source venv/bin/activate  
-3. Run in the terminal   
-pip install pytest  
+source .venv/bin/activate  
+3. Run in the terminal  
 pip install -r requirements.txt
