--- Solutions/urban_route_main_page.py
@@ -1,35 +1,0 @@
-from selenium.webdriver.common.by import By
-
-
-# Defining the page class, locators and method in the class
-class UrbanRoutesPage:
-    # Locators as class attributes
-    FROM_LOCATOR = (By.ID, 'from')
-    TO_LOCATOR = (By.ID, 'to')
-    CUSTOM_OPTION_LOCATOR = (By.XPATH, '//div[text()="Custom"]')
-    BIKE_ICON_LOCATOR = (By.XPATH, '//img[@src="/static/media/bike.fb41c762.svg"]')
-    BIKE_TEXT_LOCATOR = (By.XPATH, '//div[contains(text(),"Bike")]')
-
-    def __init__(self, driver):
-        self.driver = driver  # Initialize the driver
-
-    def enter_from_location(self, from_text):
-        # Enter From
-        self.driver.find_element(*self.FROM_LOCATOR).send_keys(from_text)
-
-    def enter_to_location(self, to_text):
-        # Enter To
-        self.driver.find_element(*self.TO_LOCATOR).send_keys(to_text)
-
-    def click_custom_option(self):
-        # Click Custom
-        self.driver.find_element(*self.CUSTOM_OPTION_LOCATOR).click()
-
-    def click_bike_icon(self):
-        # Click Bike Icon
-        self.driver.find_element(*self.BIKE_ICON_LOCATOR).click()
-
-    def get_bike_text(self):
-        # Return the "Bike" text
-        return self.driver.find_element(*self.BIKE_TEXT_LOCATOR).text
-
