+++ requirements.txt
@@ -1,1 +1,2 @@
-selenium~=4.24.0
+selenium==4.25.0
+pytest==8.3.3
